requirements.txt
torchvision
pandas
tqdm
matplotlib
torchmetrics
pytorch_toolbelt
scikit-learn
